Move DKGTable to admin folder
