+++ web/frontend/src/pages/admin/DKGTable.tsx
@@ -2,8 +2,8 @@ import usePostCall from 'components/utils/usePostCall';
 import React, { FC, useContext, useEffect, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { NodeStatus } from 'types/node';
-import DKGStatus from './DKGStatus';
-import * as endpoints from '../../../components/utils/Endpoints';
+import DKGStatus from '../../components/utils/DKGStatus';
+import * as endpoints from '../../components/utils/Endpoints';
 import { ID } from 'types/configuration';
 import { AuthContext } from 'index';
 import { UserRole } from 'types/userRole';
